Apple Retail v2/requirements.txt
pyspark
delta-spark
great-expectations
pyyaml